[Refacor Chore] accept 값 수정, type 위치 변경
